requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.3
pydantic==2.8.2
pydantic-settings==2.3.2
langchain-core==0.2.23
langchain-community==0.2.10
langchain-text-splitters==0.2.2
qdrant-client==1.9.1
requests==2.32.3
typer==0.12.3
tqdm==4.66.5
python-dotenv==1.0.1
